verification/requirements.txt
cocotb==1.7.2
cocotb-test==0.2.4
pytest-html
pytest-xdist
